rework app structure to customize clients used
