+++ client/client.go
@@ -1,6 +1,7 @@
 package client
 
 import (
+	"github.com/nats-io/nats.go"
 	"github.com/simpleiot/simpleiot/data"
 )
 
@@ -16,8 +17,27 @@ import (
 // assumes the point was generated by the client and the client already knows about it.
 // Thus, if you want points to get to a client, Origin must be set.
 type Client interface {
-	Start() error
-	Stop(error)
+	StartStop
+
 	Points(string, []data.Point)
 	EdgePoints(string, string, []data.Point)
+}
+
+// DefaultClients returns an actor for the default group of built in clients
+func DefaultClients(nc *nats.Conn) (*Group, error) {
+	g := NewGroup()
+
+	sc := NewManager(nc, NewSerialDevClient)
+	g.Add(sc)
+
+	rc := NewManager(nc, NewRuleClient)
+	g.Add(rc)
+
+	db := NewManager(nc, NewDbClient)
+	g.Add(db)
+
+	sg := NewManager(nc, NewSignalGeneratorClient)
+	g.Add(sg)
+
+	return g, nil
 }
